+++ OrderEntry/MVC/Models/StockOrderViewModel.cs
@@ -14,7 +14,7 @@
 
         public required string TOSStop { get; set; }
 
-        public required double PositionValue { get; set; }
+        public required decimal PositionValue { get; set; }
 
         public required string BackgroundColor { get; set; }
     }
