+++ OrderEntry/OrderEntry/MindfulTrader/IOrder.cs
@@ -7,10 +7,10 @@
         Strategies Strategy { get; set; }
         string Ticker { get; set; }
         int Count { get; set; }
-        double PotentialEntry { get; set; }
-        double PotentialProfit { get; set; }
-        double PotentialStop { get; set; }
-        double PositionValue { get; set; }
+        decimal PotentialEntry { get; set; }
+        decimal PotentialProfit { get; set; }
+        decimal PotentialStop { get; set; }
+        decimal PositionValue { get; set; }
         string EarningsDate { get; set; }
         string DividendsDate { get; set; }
         int EntryOrderId { get; set; }
